requirements.txt
pyspark==3.5.5
kafka-python
confluent_kafka